+++ src/pantry/commands/search.py
@@ -1,5 +1,7 @@
 """`pantry search` — one query, every provider that is willing to answer."""
 
+from decimal import Decimal
+
 import click
 from agentcli import json_option, limit_option
 
@@ -41,7 +43,7 @@ def search(
     query: tuple[str, ...],
     sources: tuple[str, ...],
     remote: bool,
-    grams: float | None,
+    grams: Decimal | None,
     json_output: bool,
     limit: int,
 ) -> None:
